+++ contracts/Vault.sol
@@ -89,72 +89,80 @@ contract Vault is IVault, ReentrancyGuardUpgradeable, OwnerPausable, BaseRelayRe
         _setTrustedForwarder(trustedForwarderArg);
     }
 
-    /// @param token The address of the token sender is going to deposit
-    /// @param amount The amount of the token sender is going to deposit
-    /// @dev The token can be other than settlementToken once multi collateral feature is implemented
-    function deposit(address token, uint256 amount) external whenNotPaused nonReentrant onlySettlementToken(token) {
+    /// @param token the address of the token to deposit;
+    ///        once multi-collateral is implemented, the token is not limited to settlementToken
+    /// @param amountX10_D the amount of the token to deposit in decimals D (D = _decimals)
+    function deposit(address token, uint256 amountX10_D)
+        external
+        whenNotPaused
+        nonReentrant
+        onlySettlementToken(token)
+    {
         address from = _msgSender();
 
-        _modifyBalance(from, token, amount.toInt256());
+        _modifyBalance(from, token, amountX10_D.toInt256());
 
-        // for deflationary token,
-        // amount may not be equal to the received amount due to the charged (and burned) transaction fee
+        // check for deflationary tokens by assuring balances before and after transferring to be the same
         uint256 balanceBefore = IERC20Metadata(token).balanceOf(from);
-        TransferHelper.safeTransferFrom(token, from, address(this), amount);
-        // balance amount inconsistent
-        require(balanceBefore.sub(IERC20Metadata(token).balanceOf(from)) == amount, "V_BAI");
+        TransferHelper.safeTransferFrom(token, from, address(this), amountX10_D);
+        // V_BAI: inconsistent balance amount
+        require(balanceBefore.sub(IERC20Metadata(token).balanceOf(from)) == amountX10_D, "V_IBA");
 
         uint256 settlementTokenBalanceCap = IClearingHouseConfig(_clearingHouseConfig).getSettlementTokenBalanceCap();
-        // greater than settlement token balance cap
+        // V_GTSTBC: greater than settlement token balance cap
         require(IERC20Metadata(token).balanceOf(address(this)) <= settlementTokenBalanceCap, "V_GTSTBC");
 
-        emit Deposited(token, from, amount);
+        emit Deposited(token, from, amountX10_D);
     }
 
-    /// @param token The address of the token sender is going to withdraw
-    /// @param amountX10_D The amount in decimal D (D = _decimals) of the token sender is going to withdraw
-    /// @dev The token can be other than settlementToken once multi collateral feature is implemented
+    /// @param token the address of the token sender is going to withdraw
+    ///        once multi-collateral is implemented, the token is not limited to settlementToken
+    /// @param amountX10_D the amount of the token to withdraw in decimals D (D = _decimals)
     function withdraw(address token, uint256 amountX10_D)
         external
         whenNotPaused
         nonReentrant
         onlySettlementToken(token)
     {
+        // the full process of withdrawal:
+        // 1. settle funding payment to owedRealizedPnl
+        // 2. collect fee to owedRealizedPnl
+        // 3. call Vault.withdraw(token, amount)
+        // 4. settle pnl to trader balance in Vault
+        // 5. transfer the amount to trader
+
         address to = _msgSender();
 
-        // the full process of a trader's withdrawal:
-        //     settle funding payment to owedRealizedPnl
-        //     collect fee to owedRealizedPnl
-        // call Vault.withdraw(token, amount)
-        //     settle pnl to trader balance in Vault
-        //     transfer amount to trader
-
-        // settle funding payments to owedRealizedPnl,
-        // while fees are ok to let maker decides whether to collect using CH.removeLiquidity(0)
+        // settle all funding payments to owedRealizedPnl
         IExchange(_exchange).settleAllFunding(to);
 
         // settle owedRealizedPnl in AccountBalance
         int256 owedRealizedPnlX10_18 = IAccountBalance(_accountBalance).settleOwedRealizedPnl(to);
 
-        // by this time free collateral should see zero funding payment and zero owedRealizedPnl
+        // by this time there should be no owedRealizedPnl nor pending funding payment in free collateral
         int256 freeCollateralByImRatioX10_D =
             getFreeCollateralByRatio(to, IClearingHouseConfig(_clearingHouseConfig).getImRatio());
         // V_NEFC: not enough freeCollateral
         require(
-            freeCollateralByImRatioX10_D.addS(owedRealizedPnlX10_18, _decimals) >= amountX10_D.toInt256(),
+            freeCollateralByImRatioX10_D.add(owedRealizedPnlX10_18.formatSettlementToken(_decimals)) >=
+                amountX10_D.toInt256(),
             "V_NEFC"
         );
 
-        // borrow settlement token from insurance fund if token balance is not enough
+        // borrow settlement token from insurance fund if the token balance in Vault is not enough
         uint256 vaultBalanceX10_D = IERC20Metadata(token).balanceOf(address(this));
         if (vaultBalanceX10_D < amountX10_D) {
-            uint256 borrowAmountX10_D = amountX10_D - vaultBalanceX10_D;
-            IInsuranceFund(_insuranceFund).borrow(borrowAmountX10_D);
-            _totalDebt += borrowAmountX10_D;
+            uint256 borrowedAmountX10_D = amountX10_D - vaultBalanceX10_D;
+            IInsuranceFund(_insuranceFund).borrow(borrowedAmountX10_D);
+            _totalDebt += borrowedAmountX10_D;
         }
 
-        // settle withdraw amount and owedRealizedPnl to collateral balance
-        _modifyBalance(to, token, (-(amountX10_D.toInt256())).addS(owedRealizedPnlX10_18, _decimals));
+        // settle withdrawn amount and owedRealizedPnl to collateral
+        _modifyBalance(
+            to,
+            token,
+            (-(amountX10_D.toInt256().sub(owedRealizedPnlX10_18.formatSettlementToken(_decimals))))
+        );
         TransferHelper.safeTransfer(token, to, amountX10_D);
 
         emit Withdrawn(token, to, amountX10_D);
@@ -170,7 +178,7 @@ contract Vault is IVault, ReentrancyGuardUpgradeable, OwnerPausable, BaseRelayRe
     }
 
     /// @inheritdoc IVault
-    /// @dev cached the settlement token's decimal for gas optimization
+    /// @dev cache the settlement token's decimals for gas optimization
     function decimals() external view override returns (uint8) {
         return _decimals;
     }
@@ -210,34 +218,39 @@ contract Vault is IVault, ReentrancyGuardUpgradeable, OwnerPausable, BaseRelayRe
     }
 
     function balanceOf(address trader) public view override returns (int256) {
-        return _getBalance(trader, _settlementToken);
+        return _balance[trader][_settlementToken];
     }
 
-    /// @param trader The address of the trader to query
-    /// @param ratio The margin requirement ratio
-    /// @dev there are three configurations for different insolvency risk tolerance: conservative, moderate, aggressive
-    /// we will start with the conservative one, then gradually change it to more aggressive ones
-    /// to increase capital efficiency.
+    /// @dev there are three configurations for different insolvency risk tolerances: conservative, moderate, aggressive
+    ///      we will start with the conservative one and gradually move to aggressive to increase capital efficiency
+    /// @param trader the address of the trader
+    /// @param ratio the margin requirement ratio, imRatio or mmRatio
+    /// @return freeCollateralByRatio freeCollateral, by using the input margin requirement ratio; can be negative
     function getFreeCollateralByRatio(address trader, uint24 ratio) public view override returns (int256) {
-        // conservative config: freeCollateral = min(collateral, accountValue) - imReq, freeCollateral could be negative
-        int256 fundingPayment = IExchange(_exchange).getAllPendingFundingPayment(trader);
-        (int256 owedRealizedPnl, int256 unrealizedPnl) =
+        // conservative config: freeCollateral = min(collateral, accountValue) - margin requirement ratio
+        int256 fundingPaymentX10_18 = IExchange(_exchange).getAllPendingFundingPayment(trader);
+        (int256 owedRealizedPnlX10_18, int256 unrealizedPnlX10_18) =
             IAccountBalance(_accountBalance).getOwedAndUnrealizedPnl(trader);
-        int256 totalCollateralValue = balanceOf(trader).addS(owedRealizedPnl.sub(fundingPayment), _decimals);
+        int256 totalCollateralValueX10_D =
+            balanceOf(trader).add(owedRealizedPnlX10_18.sub(fundingPaymentX10_18).formatSettlementToken(_decimals));
 
         // accountValue = totalCollateralValue + totalUnrealizedPnl, in the settlement token's decimals
-        int256 accountValue = totalCollateralValue.addS(unrealizedPnl, _decimals);
-        uint256 totalMarginRequirement = _getTotalMarginRequirement(trader, ratio);
-        return PerpMath.min(totalCollateralValue, accountValue).subS(totalMarginRequirement.toInt256(), _decimals);
+        int256 accountValueX10_D = totalCollateralValueX10_D.add(unrealizedPnlX10_18.formatSettlementToken(_decimals));
+        uint256 totalMarginRequirementX10_18 = _getTotalMarginRequirement(trader, ratio);
 
-        // moderate config: freeCollateral = max(min(collateral, accountValue - imReq), 0)
-        // return PerpMath.max(PerpMath.min(collateralValue, accountValue.subS(totalImReq, decimals)), 0).toUint256();
+        return
+            PerpMath.min(totalCollateralValueX10_D, accountValueX10_D).sub(
+                totalMarginRequirementX10_18.toInt256().formatSettlementToken(_decimals)
+            );
 
-        // aggressive config: freeCollateral = max(accountValue - imReq, 0)
-        // TODO note that aggressive model depends entirely on unrealizedPnl, which depends on the index price, for
-        //  calculating freeCollateral. We should implement some sort of safety check before using this model;
-        //  otherwise a trader could drain the entire vault if the index price deviates significantly.
-        // return PerpMath.max(accountValue.subS(totalImReq, decimals), 0).toUint256()
+        // moderate config: freeCollateral = min(collateral, accountValue - imReq)
+        // return PerpMath.min(collateralValue, accountValue.subS(totalImReq.formatSettlementToken(_decimals));
+
+        // aggressive config: freeCollateral = accountValue - imReq
+        // note that the aggressive model depends entirely on unrealizedPnl, which depends on the index price
+        //      we should implement some sort of safety check before using this model; otherwise,
+        //      a trader could drain the entire vault if the index price deviates significantly.
+        // return accountValue.subS(totalImReq.formatSettlementToken(_decimals));
     }
 
     //
@@ -249,7 +262,7 @@ contract Vault is IVault, ReentrancyGuardUpgradeable, OwnerPausable, BaseRelayRe
         address token,
         int256 amount
     ) internal {
-        _balance[trader][token] = _getBalance(trader, token).add(amount);
+        _balance[trader][token] = _balance[trader][token].add(amount);
     }
 
     //
@@ -260,10 +273,6 @@ contract Vault is IVault, ReentrancyGuardUpgradeable, OwnerPausable, BaseRelayRe
     function _getTotalMarginRequirement(address trader, uint24 ratio) internal view returns (uint256) {
         uint256 totalDebtValue = IAccountBalance(_accountBalance).getTotalDebtValue(trader);
         return totalDebtValue.mulRatio(ratio);
-    }
-
-    function _getBalance(address trader, address token) internal view returns (int256) {
-        return _balance[trader][token];
     }
 
     /// @inheritdoc BaseRelayRecipient