+++ contracts/lib/UniswapV3Broker.sol
@@ -78,6 +78,8 @@ library UniswapV3Broker {
         uint256 quote;
     }
 
+    uint256 internal constant _DUST = 10;
+
     function addLiquidity(AddLiquidityParams memory params) internal returns (AddLiquidityResponse memory) {
         // zero inputs
         require(params.base > 0 || params.quote > 0, "UB_ZIs");
@@ -154,7 +156,13 @@ library UniswapV3Broker {
 
         // if no price limit, require the full output amount as it's technically possible for amounts to not match
         // UB_UOA: unmatched output amount
-        if (!params.isExactInput && params.sqrtPriceLimitX96 == 0) require(exactAmount == params.amount, "UB_UOA");
+        if (!params.isExactInput && params.sqrtPriceLimitX96 == 0) {
+            require(
+                (exactAmount > params.amount ? exactAmount.sub(params.amount) : params.amount.sub(exactAmount)) < _DUST,
+                "UB_UOA"
+            );
+            return params.isBaseToQuote ? SwapResponse(amount0, params.amount) : SwapResponse(params.amount, amount1);
+        }
 
         return SwapResponse(amount0, amount1);
     }